[P24-19] feat: 빈값 예외수정 및 SecurityConfig 파일위치 이동
- NotBlank 관련 예외 수정
- SecurityConfig 파일 위치 이동
